Refactor code to implement ability to display data of entered city name
This commit allows the user to enter a valid city name in which the program will display the correct temperatures at the current active scale. It utilizes a pre-existing module with slight refactoring that now bases on two conditions: (1) a city name to fetch the corresponding data and (2) the active temperature scale. This allows the module to return any set of data based on the given parameters as opposed to the default value ('seattle').